+++ comicbox/cli.py
@@ -108,6 +108,12 @@ def _add_option_group(parser):
         " keyC: {subkey: {subsubkey: value}') Place a space after colons so they are properly parsed as YAML key value pairs.",
     )
     option_group.add_argument(
+        "-D",
+        "--delete-keys",
+        action=CSVAction,
+        help="Delete a comma delimited list of comicbox keys entirely from the final metadata.",
+    )
+    option_group.add_argument(
         "-d",
         "--dest-path",
         help="destination path for extracting pages and metadata.",
@@ -135,6 +141,13 @@ def _add_option_group(parser):
         action="store_false",
         default=True,
         help=("Never compute page_count or pages metadata from the archive."),
+    )
+    option_group.add_argument(
+        "-R",
+        "--replace-metadata",
+        action="store_true",
+        default=False,
+        help="Replace metadata keys instead of merging them.",
     )
     option_group.add_argument(
         "-Q",