+++ NEWS.md
@@ -1,5 +1,12 @@
 # 📰 News
 
+## v1.2.0
+
+- Add -R --replace_metadata option. Default behavior is to merge keys.
+- Add -D --delete_keys option.
+- Faster metadata writing. Replace files in zipfiles instead of rewriting the
+  entire archive.
+
 ## v1.1.10
 
 - Fix parsing negative issue numbers in filenames.